update and refactor news components
