+++ src/common/containers/news/article-info/styles.scss
@@ -1,0 +1,20 @@
+.news-article-info-container {
+  gap: 10px;
+  width: auto;
+  padding: 2px 0px;
+
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: row;
+  align-items: center;
+
+  .dot-container {
+    gap: 10px;
+    display: flex;
+    align-items: center;
+  }
+
+  .grey {
+    color: #555656;
+  }
+}
