+++ src/common/containers/news/author-info/styles.scss
@@ -1,0 +1,7 @@
+.news-author-container {
+  gap: 8px;
+  width: auto;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
