+++ src/common/containers/news/index.js
@@ -1,0 +1,3 @@
+export { NewsGrid } from "./news-grid";
+export { NewsAuthorInfo } from "./author-info";
+export { NewsArticleInfo } from "./article-info";
