+++ src/common/containers/news/news-grid/styles.scss
